Move seed init to the top level
